+++ src/ezdxf/addons/drawing/backend.py
@@ -72,17 +72,6 @@ class BackendInterface(ABC):
     def finalize(self) -> None:
         raise NotImplementedError
 
-    @abstractmethod
-    def set_clipping_path(
-        self, path: Optional[Path] = None, scale: float = 1.0
-    ) -> bool:
-        """Set the current clipping path.
-        Returns True if a clipping path is supported.
-        An empty path or None removes the clipping path.
-        The `scale` is the scaling factor from modelspace to viewport.
-        """
-        raise NotImplementedError
-
 
 class Backend(BackendInterface, metaclass=ABCMeta):
     def __init__(self) -> None:
@@ -107,12 +96,6 @@ class Backend(BackendInterface, metaclass=ABCMeta):
     @abstractmethod
     def set_background(self, color: Color) -> None:
         raise NotImplementedError
-
-    def set_clipping_path(
-        self, path: Optional[Path] = None, scale: float = 1.0
-    ) -> bool:
-        """Clipping path is not supported by default."""
-        return False
 
     @abstractmethod
     def draw_point(self, pos: AnyVec, properties: Properties) -> None: