+++ NEWS.md
@@ -13,6 +13,7 @@ Version 1.0.4b1 - dev
 - CHANGED: text rendering for the `drawing` add-on and text measurement is done by the
   `fontTools` package
 - CHANGED: moved text rendering from backend classes to the `Frontend` class
+- CHANGED: move clipping support from backend classes to the `Frontend` class
 - REMOVED: `PillowBackend` and the `pillow` command
 - REMOVED: `geomdl` dependency
 - BUGFIX: invalid bulge to Bezier curve conversion for bulge values >= 1